Add websocket based Ambeeance client
This is quite rough around the edges, could definitely use some error
handling/disconnect recovery/etc. but you can give it an object and
its flat keys will be sent to Ambeeance and autospec'd.

Changes to values in Ambeeance will be assigned to the object if it
has a matching key.